--- Tests/Realm.Tests/Sync/TestObjects.cs
@@ -1,63 +1,0 @@
-﻿////////////////////////////////////////////////////////////////////////////
-//
-// Copyright 2016 Realm Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-////////////////////////////////////////////////////////////////////////////
-
-using MongoDB.Bson;
-
-namespace Realms.Tests.Sync
-{
-    public class HugeSyncObject : RealmObject
-    {
-        [PrimaryKey]
-        [MapTo("_id")]
-        public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
-
-        public byte[] Data { get; set; }
-
-        public HugeSyncObject()
-        {
-        }
-
-        public HugeSyncObject(int dataSize)
-        {
-            var data = new byte[dataSize];
-            TestHelpers.Random.NextBytes(data);
-            Data = data;
-        }
-    }
-
-    [Explicit]
-    public class HugeSyncAsymmetricObject : AsymmetricObject
-    {
-        [PrimaryKey]
-        [MapTo("_id")]
-        public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
-
-        public byte[] Data { get; set; }
-
-        public HugeSyncAsymmetricObject()
-        {
-        }
-
-        public HugeSyncAsymmetricObject(int dataSize)
-        {
-            var data = new byte[dataSize];
-            TestHelpers.Random.NextBytes(data);
-            Data = data;
-        }
-    }
-}
